# contracts/v1/openapi.json
{
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Loopback-only local recommendations for the PLVA core. schema_version 1.0. The core enforces every decision; this service only recommends.",
    "title": "PLVA private reasoning service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/health": {
      "get": {
        "responses": {
          "200": {
            "description": "{\"status\": \"alive\"}"
          }
        },
        "summary": "Process liveness only. Unauthenticated."
      }
    },
    "/v1/approve": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./approve-request.schema.json"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./approve-response.schema.json"
                }
              }
            },
            "description": "Typed recommendation; denials and errors are ordinary results."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./error-body.schema.json"
                }
              }
            },
            "description": "SCHEMA_INVALID: body rejected; input is never echoed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./error-body.schema.json"
                }
              }
            },
            "description": "UNAUTHORIZED"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./error-body.schema.json"
                }
              }
            },
            "description": "FORBIDDEN_HOST or FORBIDDEN_ORIGIN"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./error-body.schema.json"
                }
              }
            },
            "description": "DUPLICATE_REQUEST: (session_id, request_id) already seen"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./error-body.schema.json"
                }
              }
            },
            "description": "PAYLOAD_TOO_LARGE"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./error-body.schema.json"
                }
              }
            },
            "description": "UNSUPPORTED_MEDIA_TYPE"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./error-body.schema.json"
                }
              }
            },
            "description": "NOT_READY"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Recommend approve/deny for one private action."
      }
    },
    "/v1/compute": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./compute-request.schema.json"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./compute-response.schema.json"
                }
              }
            },
            "description": "Typed recommendation; denials and errors are ordinary results."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./error-body.schema.json"
                }
              }
            },
            "description": "SCHEMA_INVALID: body rejected; input is never echoed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./error-body.schema.json"
                }
              }
            },
            "description": "UNAUTHORIZED"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./error-body.schema.json"
                }
              }
            },
            "description": "FORBIDDEN_HOST or FORBIDDEN_ORIGIN"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./error-body.schema.json"
                }
              }
            },
            "description": "DUPLICATE_REQUEST: (session_id, request_id) already seen"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./error-body.schema.json"
                }
              }
            },
            "description": "PAYLOAD_TOO_LARGE"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./error-body.schema.json"
                }
              }
            },
            "description": "UNSUPPORTED_MEDIA_TYPE"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./error-body.schema.json"
                }
              }
            },
            "description": "NOT_READY"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Bounded sort/select; returns input tokens only."
      }
    },
    "/v1/readiness": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./readiness-response.schema.json"
                }
              }
            },
            "description": "Readiness"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Authenticated readiness, mode, isolation state, capabilities."
      }
    },
    "/v1/review-trace": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./review-trace-request.schema.json"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./review-trace-response.schema.json"
                }
              }
            },
            "description": "Typed recommendation; denials and errors are ordinary results."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./error-body.schema.json"
                }
              }
            },
            "description": "SCHEMA_INVALID: body rejected; input is never echoed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./error-body.schema.json"
                }
              }
            },
            "description": "UNAUTHORIZED"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./error-body.schema.json"
                }
              }
            },
            "description": "FORBIDDEN_HOST or FORBIDDEN_ORIGIN"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./error-body.schema.json"
                }
              }
            },
            "description": "DUPLICATE_REQUEST: (session_id, request_id) already seen"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./error-body.schema.json"
                }
              }
            },
            "description": "PAYLOAD_TOO_LARGE"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./error-body.schema.json"
                }
              }
            },
            "description": "UNSUPPORTED_MEDIA_TYPE"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./error-body.schema.json"
                }
              }
            },
            "description": "NOT_READY"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Recommend continue/warn/halt over a value-free trace."
      }
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:18555"
    }
  ]
}